+++ sample/src/main/java/com/clevertap/demo/MyApplication.kt
@@ -5,7 +5,9 @@ import android.app.Application.ActivityLifecycleCallbacks
 import android.app.NotificationManager
 import android.content.Context
 import android.content.Intent
+
 import android.os.Build
+
 import android.os.Bundle
 import android.util.Log
 import androidx.multidex.MultiDex